Sensitivity for vecnorm
Implemented the sentivity for `vecnorm` and fixed up the unit tests to
make them work with `vecnorm`. Specifically, I removed the 7th argument
from `binary_linalg_optimisations`, since that was a leftover from
previous work, and I reworked the unit tests for `linalg/generic.jl` to
generate random quantities that have shapes consistent with the
types of the arguments of the function that is being tested.