Higher-order finite difference methods (#60)
* Finite-difference methods of arbitrary order

* Uncommented commented unit tests

* Fixed formatting error in docstring of fdm

* Reduced tolerance of array.jl unit tests

* Changed unit tests to use a central 5/1 method

Also changed the check for approximate numerical equality. We now have
the theoretical guarantee that large numbers will be within a factor 2
of the tolerance on their relative difference, and that small numbers
will also be within a factor 2 of the tolrance on their absolute
difference.

* Let fdm and related functions use nabla types

* Small changes in documentation

* Reduced order of methods in units tests of fdm

A 14th order method cant be constructed on 32 bit machines due to the
fact that `factorial(14)` generates an overflow on 32 bit machines. I
therefore also removed the restriction that the order of the method must
be less than 20, since that only applies to 64 bit systems.

* Added a test for FDMReport and printing thereof

* Explicity check that an FDM can be computed

* Fixed order of FDM in unit test

* Tightened and standardised unit tests

* Improvements to unit test framework

- Separated out a function that checks for approximate numerical
  equality. This function throws a nice error message in the case of
  failure.
- Unit test for checking for approximate numerical equality.
- Simplified and capitalised the names of test cases.

* Made unit tests for fdm easier to pass