DiffRules integration (#75)
* added functions to determine input domains

* implemented scalar sensitivities with DiffRules

* added missing variables

* removed commented code and refactored scalar tests

* added DiffRules to REQUIRE

* increased upper bound in fdms + custom beta domain

Some tests were failing due to very large derivatives. This commit hopes
to fix that by increasing the assumed upper bound in the FDMs. Also, it
turns out that `beta` accepts negative arguments, instead of throwing a
`DomainError`. To address that, this commit implements a specialised
version for `domain2`.

* added test for beta's domain

* reformatting and tolerance tweaks

* explicit import of DiffRule

This hopefully fixes the CI build.

* removed needs_output functionality

This is not necessary anymore due to the DiffRules integration. Also removed
commented out code.

* cleaned up tests

- Removed commented out code.
- Made tests deterministic.
- Slightly adjusted the configurion of the FDMs.