Add line information to more generated code
In some places we are willfully omitting line information by passing
`nothing` to certain `Expr`s, and in others we're generating expressions
that are syntactically valid but are not what Julia itself produces.
Part of that is the omission of line information.

This changes the function definition expressions from things like
```julia
f(x::Node{Int}) = Branch(f, (x,), getfield(x, :tape))
```
to
```julia
function f(x::Node{Int})
    #= line info here =#
    Branch(f, (x,), getfield(x, :tape))
end
```
"What's wrong with the former?" you might ask. Nothing in principle,
though if you were to write that in your own code, Julia parses it as
```julia
f(x::Node{Int}) = begin
    #= line info here =#
    Branch(f, (x,), getfield(x, :tape))
end
```
For ease of visual parsing, we'll change it to use `:function`
expression heads, i.e. long-form function definitions, as it's
functionally equivalent.

The benefit of having line information is for backtraces. If something
goes wrong, the user should now be able to get a better idea of where a
problem is, rather than Julia providing a backtrace that just says
something along the lines of "in Nabla."