Make printing of Tapes more verbose (and informative)
Currently the printing of `Tape`s leaves a bit to be desired in terms of
examining the actual contents, due in part to the lack of separation
between position and value, and due to how values are printed. With this
change, values are printed using `show`, which is the intended way to
display things nicely, and positions are wrapped in brackets with a
colon. Additionally, a header is printed that lists the number of tape
elements, in the same manner as e.g. `Dict`s.