Use explicit_intercepts for sum (#146)
This allows us to consume and record `sum` calls as such, rather than
simply recording that `mapreduce` was called. This should make reading
the tape a bit more informative, plus it permits specializations on the
particular mapping function `f` passed in `sum(f, x)`.

Because of the aforementioned specialization feature, we can now add a
specialization for `sum(abs2, x)`, which is a common pattern and is
equivalent to `dot(x, x)` and `sum(x.^2)`.