Fix map extension for Nodes (#151)
Consider the function call `map(f, xs...)`. We want to use Nabla if any
of the `xs` are `Node`s. Unfortunately that isn't expressible in a
coherent way in Julia after 0.6. What we currently do introduces method
ambiguities, so we overwrite the Base method for `map` on numeric arrays
to call `invoke`. That workaround itself causes other ambiguities for
packages which also extend `map`, e.g. StaticArrays, as observed in
issue #136.

The solution here is to limit what we seek to detect among the
aforementioned `xs`. If we limit ourselves to, say, the first 10 iterable
arguments to `map` for extension with `Node`s, everything is much
simpler and we don't need to do any method overwriting. As currently
implemented, the cutoff is 10, though it can be easily changed to any
number; 10 just seemed a reasonable enough cutoff.

Fixes #136.