Fix FDM version bounds, bump Nabla version to 0.11.1 (#157)
Currently FDM's version is restricted to 0.1.0 or 0.2.0, which doesn't
really make sense; any version of FDM since 0.1.0 should be fully
compatible with Nabla. Imposing the version restriction makes later
versions of FDM uninstallable in the global environment if Nabla was
installed first.

This change sets FDM to have a lower bound only, it frees DualNumbers to
have a lower bound only rather than a fixed version, it adds an explicit
caret to the Julia bound (non-functional change as carets are implicit),
and it sets Nabla's own version to 0.11.1 in preparation for a new tag.