Test transpose wrappers for generic linalg sensitivities
Currently, when passed `Transpose` or `Adjoint` as a type, the test-only
`trand*` functions return a regular matrix. Because of this, we were not
testing the sensitivity definitions specific to those types, but rather
we were just testing the `AbstractMatrix` definitions a bunch of times.

This change ensures that we're actually hitting those code paths.