Use named tuples for more reliable dispatch in SVD (#163)
Currently we're assuming that a `getproperty` call for `SVD` objects
will return a vector for `S`, an `Adjoint` for `V`, and any other kind
of matrix for `U`. This is a bit fragile and requires making a copy for
`V` to ensure it's an `Adjoint`.

We can do better! If we have `getproperty`'s sensitivity return a
singleton `NamedTuple` containing only the requested property, we can be
sure of exactly which property `svd`'s sensitivity is getting.