Refactor generic linear algebra tests (#167)
This refactors the generic linear algebra tests to use test sets, each
with a separate RNG state. This avoids issues where running a particular
group of tests multiple times fails, apparently due to the RNG state.

A particular RNG seed was chosen for the unary tests that allows them
not to fail, and seeds for other tests were chosen arbitrarily.

This commit also has a (somewhat) unrelated change that simplifies the
sensitivity definition for `-` from `map`ping `-` to just applying it.