Simplify working with keyword arguments
Currently, `@explicit_intercepts` generates a new function when given
keyword arguments. It uses a function called `kwfname` to determine what
the name of this new function should be. This is kind of weird and
convoluted. Instead, we can just store the keyword arguments used for
the call as part of the `Branch` type, and make sure they're passed
around as appropriate. This allows us dump `kwfname` and all of that
entirely.